Test Flashing & Partial Interface Filling
Commit handling a small flashing image in the XAML of the new Map Interface, which was added to both shorten the Main Window XAML and to simplify a few more processes.

Added external color picker into the project and roughly bound it to the province map. It isn't perfect and needs some tuning, and it currently works great.

Fixed handling of CSV provinces that had periods "." in their colors. Program is well on its way to its first release.

 Handling of Cores and StateBuildings is needed before save function is implemented. Selected province highlight and a few mod-compatability errors should be relatively easy to fix.